Agent Development Kit Demo/POC/pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.7.0",
    "duckduckgo-search==8.1.1",
    "google-adk>=1.17.0",
    "langchain-community>=0.4.1",
    "litellm>=1.79.1",
]

[tool.uv.pip]
upgrade = true

[tool.ruff]
ignore = ["D"] # Ignores all 'D' rules (docstring-related)

[build-system]
requires = ["uv"]
build-backend = "uv_build"

[tool.uv.extra-build-dependencies]
poc = ["uv_build"]

[tool.uv.build-backend]
namespace = true


[project.scripts]
agent_team = "agent_team.agent:main"
stock_agent = "stock_advisor_workflow.agent:run_stock_advisor_workflow_sync"
weather_agent = "weather_time_tool_agent.agent:run_stock_advisor_workflow_sync"
stateful_agent_team = "agent_team.stateful_agents:main"